Innitial commit

+++ Exercise_repaso/exercise4.py
@@ -1,4 +1,4 @@
-numero = int(input("Ingresa un número entero: "))
+numero = float(input("Ingresa un número entero: "))
 
 if numero % 2 == 0: #el % funciona como divisor y el 2 como dividendo, el numero ingresado se divide por 2 y el programa muestra automaticamente si es par o impar
     print("El número es par.")
